Rename hubot-stonks to hubot-stonk-checker
Turns out, I didn't check on the name and it's already taken by a
project that apparently has one purpose -- to use the name I wanted.
It's not a real project and you can't even install it, but it's
squatting on the namespace on npmjs.org and that's just mean.