Adjust output for rich text vs plain text.
This commit ditches the cards/rich text specifically for slack and goes
back to emojs and text. It also now detects if this module is being used
via the slack adapter (and thus shows emoji) or if not, just uses plain
text.

I assume chat platforms other than slack allow for emoji, so we might
have to adjust the logic to determine if `rt` or "rich text" is enabled
in the future.